// knowledge/types/OutputSchema.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface OutputSchema {
  name_id: string;
  name_pretty: string;
  releases: {
    version: string;
    platforms: {
      arch: "amd64" | "x86" | "arm64" | "armv7l" | "armv6" | "ppc64" | "ppc64le" | "s390x" | "riscv64";
      os: "linux" | "freebsd" | "openbsd" | "netbsd" | "aix" | "solaris" | "darwin" | "windows";
      download_url?: string;
      download_sha256sum?: string;
      download_size?: string;
    }[];
  }[];
}